// src/mastra/tools/api/zodSchema.ts
export const strapiTemplateSchema = z.object({
    background: z.string().optional(),
    description: z.string(),
    how: z.array(z.object({
        description: z.string(),
        icon: z.string(),
        title: z.string()
    })).optional(),
    image: z.string().optional(),
    output_image: z.string().optional(),
    pre_title: z.string(),
    price: z.object({
        is_strikethrough: z.number().optional().default(0),
        previous_price: z.string().optional(),
        price: z.string()
    }).optional(),
    requirements: z.object({
        description: z.string(),
        list: z.array(z.string())
    }).optional(),
    tags: z.array(z.object({
        icon: z.string(),
        text: z.string()
    })),
    title: z.string(),
    what: z.object({
        description: z.string(),
        goal: z.string()
    }).optional(),
    why: z.object({
        advantages: z.array(z.string()),
        reasons: z.array(z.object({
            description: z.string(),
            icon: z.string(),
            title: z.string()
        }))
    }).optional()
});

export const cpReqTemplateSchema = z.object({
    category_id: z.number(),
    config: z.string(),
    description: z.string().optional(),
    id: z.number(),
    name: z.string(),
    price: z.string().optional(),
    strapi: strapiTemplateSchema.optional(),
    workspace_id: z.number().optional()
});